1. Convert Long data type of projectId to Integer 2. Convert Integer data type of durationMinutes to Short 3. Fetch employee Timesheet requests as per manager ifd and client filter

+++ src/main/java/com/apmosys/employeeportal/dto/TimesheetDTO_new/ActivityTimesheetDTO.java
@@ -56,7 +56,7 @@ public class ActivityTimesheetDTO {
      * Required - must be greater than 0
      */
     @NotNull("Duration minutes is required")
-    private Integer durationMinutes;
+    private Short durationMinutes;
 
    
 }
